requirements.txt
playwright>=1.40
beautifulsoup4>=4.12
lxml>=5.0
python-jobspy>=1.1
requests>=2.31
pandas>=2.0
flask>=3.0
